+++ Warehouse.php
@@ -233,6 +233,14 @@ if ( ROSARIO_DEBUG )
 	// @since 5.0 Load Kint.
 	Kint();
 }
+else
+{
+	function d()
+	{
+		// Prevent PHP Fatal error if Kint debug d() function not loaded.
+		return var_dump( func_get_args() );
+	}
+}
 
 
 /**